Move player event handlers to PlayerListener
